Reduce node allocations (by count, not bytes)
Use C99's flexible array member to combine the node's allocation with
the its position allocation.

This isn't that important on it's own, but it's a good setup for
exposing the node creation function & allowing the caller to perform the
allocation.

Oddly enough, I'm seeing another ~5% perf improvement when performing
lots of searches with small-ish trees.  I expected to see only a slight
speedup on the creation of the trees, but I'm not.  It's possible that
the search improvement is due to re-ordering the struct members.