Remove hyperrect allocations (count, not bytes)
Make a single allocation for the full tree struct (not the nodes) and
place the hyperrects at offsets within the allocation.  This isn't too
important for performance, but it sets us up very nicely to allow the
caller to provide a single memory buffer for the tree struct.

Now, there's only 2 main allocations - the tree struct and each node.
We should be able to create a few public functions to allow the caller
to control these without having to override malloc.