Remove _f node insert overloads
These add some overhead to the API, but the real reason for removal is
to change the internal format for coordinates from doubles -> floats.
In order to make this configurable, I plan on using a #define to allow
users to switch it at compile time.  This is going to be a lot easier
without the overloads.