De-duplicate & optimize zero weight node filter
I'm trying to make changes to our custom zero weight filter, but I don't
want to have to do it in multiple places.

This also seems to be a very minor optimization.  It may not be
statistically significant, but I thought I'd mention it since a few
other approaches that I tried seemed to have a small performance
regression.  A quick inspection of GCC's assembly output indicates that
it is already vectorizing the distance computation, so I'm guessing our
filter interrupted that somehow when doing both at the same time.