Make the zero weight filter explicit
Hiding this filter in the implementation was an un-intuitive design
choice that only makes sense when considering the single use case inside
TestFit.  That's not really appropriate for a module that is independent
and presents itself as generic.

Furthermore, since this is now a parameter, the caller can disable it to
get standard kdtree behavior.